Cleaned up handling of the workDir argument.  It got an error in some cases if it was specified as a relative path. Also made sure that the specified directory wasn't deleted if it previously existed.  This could be disasterous if it was specified as "." for example.
git-svn-id: svn://svn.r-forge.r-project.org/svnroot/odfweave@362 edb9625f-4e0d-4859-8d74-9fd3b1da38cb

+++ pkg/odfWeave/R/odfWeave.R
@@ -13,11 +13,7 @@ function(file, dest, workDir=odfTmpDir(), control=odfWeaveControl())
    zipCmd <- control$zipCmd
    zipCmd <- gsub("$$file$$", shellQuote(basename(file)), zipCmd, fixed=TRUE)
    
-   currentLocale <- c(Sys.getlocale("LC_CTYPE"), Sys.getlocale("LC_COLLATE"))
-   #Sys.setlocale("LC_CTYPE", "C")
-   #Sys.setlocale("LC_COLLATE", "C")
-
-   if(dirname(dest) == ".") dest <- paste(currentLoc, "/", dest, sep = "")
+   dest <- canonicalFilePath(dest)
    
    verbose <- control$verbose
 
@@ -42,13 +38,22 @@ function(file, dest, workDir=odfTmpDir(), control=odfWeaveControl())
       }
    }
 
-   # create temp dir to work in
+   # Create temp dir to work in.  Remember if we created it so
+   # we will know whether to delete it or not.
    if(!file.exists(workDir))
    {
       announce(verbose, "  Creating ", workDir, "\n")
       dir.create(workDir, showWarnings = TRUE, recursive = FALSE)
       if(!file.exists(workDir)) stop("Error creating working directory")
+      created.workDir <- TRUE
+   } else {
+      warning("working in existing directory")
+      created.workDir <- FALSE
    }
+
+   # Make absolute version of workDir so it will be valid after
+   # we cd to it
+   workDir <- canonicalDirPath(workDir)
 
    workingCopy <- basename(file)
    
@@ -131,9 +136,6 @@ function(file, dest, workDir=odfTmpDir(), control=odfWeaveControl())
    # Sweave results to new xml file
    announce(verbose, "  Sweaving ", rnwFileName, "\n\n")
 
-   #Sys.setlocale("LC_CTYPE", currentLocale[1])
-   #Sys.setlocale("LC_COLLATE", currentLocale[2])
-
    Sweave(file=rnwFileName, output="content_1.xml",
       quiet=!control$verbose, driver=RweaveOdf(), control=control,
       encoding="UTF-8")
@@ -141,9 +143,6 @@ function(file, dest, workDir=odfTmpDir(), control=odfWeaveControl())
    # reset the figure captions
    
    .odfEnv$fig.caption <- NULL
-
-   #Sys.setlocale("LC_CTYPE", "C")
-   #Sys.setlocale("LC_COLLATE", "C")
 
    if (!control$debug)
    {
@@ -249,9 +248,6 @@ function(file, dest, workDir=odfTmpDir(), control=odfWeaveControl())
    announce(verbose, "  Resetting wd\n")
    setwd(currentLoc)
 
-   #Sys.setlocale("LC_CTYPE", currentLocale[1])
-   #Sys.setlocale("LC_COLLATE", currentLocale[2])
-
    assign(
       "picPath",
       NA,
@@ -261,8 +257,8 @@ function(file, dest, workDir=odfTmpDir(), control=odfWeaveControl())
       NA,
       pos = .odfEnv)
 
-   # delete tmp dir that we were working in
-   if(control$cleanup)
+   # Delete tmp dir that we were working in if we created it
+   if(control$cleanup && created.workDir)
    {
       announce(verbose, "  Removing ", workDir, "\n")
       unlink(workDir, recursive=TRUE)
@@ -275,6 +271,30 @@ function(file, dest, workDir=odfTmpDir(), control=odfWeaveControl())
    }
    announce(verbose, "\n  Done\n")   
    invisible(NULL)
+}
+
+# This works like normalizePath except that it handles the case
+# where the path doesn't exist.  The value of dirname(path) must
+# exist, however.
+"canonicalFilePath" <- function(path)
+{
+   d <- dirname(path)
+   if (! file.exists(d))
+      stop('directory does not exist: ', d)
+   if (! file_test("-d", d))
+      stop(d, ' is not a directory')
+   file.path(normalizePath(d), basename(path))
+}
+
+# This is a wrapper for normalizePath, but issues specific
+# errors rather than a warning if the path doesn't exist.
+"canonicalDirPath" <- function(path)
+{
+   if (! file.exists(path))
+      stop('directory does not exist: ', path)
+   if (! file_test("-d", path))
+      stop(path, ' is not a directory')
+   normalizePath(path)
 }
 
 "announce" <- function (verbose = TRUE, ...)
